consis the new shard strings test (#1442)
